+++ docs/pages/documentation/general/GeneralNav.vue
@@ -6,6 +6,7 @@
         <router-link to="modal" class="navbar-item is-tab">Modal</router-link>
         <router-link to="panel" class="navbar-item is-tab">Panel</router-link>
         <router-link to="tabs" class="navbar-item is-tab">Tabs</router-link>
+        <router-link to="tag" class="navbar-item is-tab">Tag</router-link>
         <router-link to="tooltip" class="navbar-item is-tab">Tooltip</router-link>
     </div>
 </template>